Restructure database and mail modules to dedicated internal folders
